Allow tooltips to break out of jail
If a calendar is placed in a div that has overflow:hidden, the tooltips
would be clipped. By creating a separate div at the end of the document
body to hold the current tooltip, the overflow property of the calendar
container does not affect the display of the tooltip.

Tooltip position calculation was also adjusted so that the tooltip will
display within the current viewport (and switch sides if it extends off
the left side of the viewport) as long as the tooltip size does not
exceed the size of the viewport.